% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_table_snail2}
\alias{codon_table_snail2}
\title{Codon table snail 2}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_table_snail2
}
\description{
Here we try making a codon table
}
\keyword{datasets}
