% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3.R
\name{divide}
\alias{divide}
\title{Title}
\usage{
divide(input, start = 1)
}
\arguments{
\item{input}{string of bases}

\item{start}{first base in the string of bases}

\item{divide}{this is basically a function that returns codons from the string of bases}
}
\value{
vector of codons
}
\description{
Title
}
