% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran_dna.R
\name{ran_dna}
\alias{ran_dna}
\title{Title}
\usage{
ran_dna(c_length)
}
\arguments{
\item{c_length}{the length of the codon vector you would like}
}
\value{
returns a randomly generated sequence of nucleotides based on the number you entered
}
\description{
Title
}
