% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lians.R
\name{lians}
\alias{lians}
\title{Title}
\usage{
lians(c_length)
}
\arguments{
\item{c_length}{the length of the codon vector you would like}
}
\value{
returns a list of codons based on the number you entered
}
\description{
Title
}
