% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function5.R
\name{plot_function5}
\alias{plot_function5}
\title{Plot_function5}
\usage{
plot_function5(input)
}
\arguments{
\item{input}{is character string of nucleotide sequence or amino acid sequence}
}
\value{
the output is a plot that shows the number of times each nucleotide occurs
}
\description{
Plot_function5
}
\examples{
x <- "AUUUUUCTCT"
plot_function5(x)
}
