% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func1_ran_dna.R
\name{ran_dna}
\alias{ran_dna}
\title{Create random dna sequences of defined length}
\usage{
ran_dna(seq_length)
}
\arguments{
\item{seq_length}{the length of the codon vector you would like}
}
\value{
returns a list of codons based on the number you entered
}
\description{
Create random dna sequences of defined length
}
