% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut_function2.R
\name{ut_function2}
\alias{ut_function2}
\title{ut_function2}
\usage{
ut_function2(dna_seq)
}
\arguments{
\item{dna_seq}{is a character string of nucleotides (DNA) where T will be replaced with U}
}
\value{
gives us the modified UT string (RNA) where every T is replaced by U
}
\description{
ut_function2
}
\examples{
x <- "ATT"
ut_function2(x)
}
